+++ Assets/Scripts/HealthBar.cs
@@ -11,12 +11,12 @@ public class HealthBar : MonoBehaviour
     {
         //slider = GetComponent<Slider>();
     }
-    public void SetMaxHealth(int maxHealth) 
+    public void SetMaxHealth(float maxHealth) 
     {
         slider.maxValue = maxHealth;
         slider.value = maxHealth;
     }
-    public void SetCurrentHealth(int currHealth) 
+    public void SetCurrentHealth(float currHealth) 
     {
         slider.value = currHealth;
     }
